requirements.txt
# Este arquivo é apenas para referência
# No NixOS, os pacotes são gerenciados pelo shell.nix
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly==5.18.0
pydantic==2.5.0